+++ tests/testthat/test-url.R
@@ -82,10 +82,13 @@ test_that("handles all atomic vectors", {
   expect_equal(format_query_param(" "), "%20")
 })
 
+test_that("doesn't add extra spaces", {
+  expect_equal(format_query_param(c(1, 1000)), c("1", "1000"))
+  expect_equal(format_query_param(c("a", "bcdef")), c("a", "bcdef"))
+})
 
 test_that("formats numbers nicely", {
   expect_equal(format_query_param(1e9), "1000000000")
-  expect_equal(format_query_param(c(1, 1000)), c("1", "1000"))
 })
 
 test_that("can opt out of escaping", {